+++ lib/components/button/hm_button.dart
@@ -1,6 +1,8 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_hooks/flutter_hooks.dart';
 import 'package:hmwidget/size/hm_button_size.dart';
+import 'package:hmwidget/utils/constant.dart';
+import 'package:hmwidget/utils/helper.dart';
 import 'package:hmwidget/utils/hm_raduis.dart';
 import 'package:styled_widget/styled_widget.dart';
 
@@ -23,7 +25,9 @@ class HMButton extends HookWidget {
     required this.onPressed,
     this.disabled = false,
     this.hidden = false,
-  }) : assert(icon == null && iconAtLeft == null);
+  })  : assert(icon == null && iconAtLeft == null),
+        assert(buttonVariant == HMButtonVariant.filled || textColor != null,
+            "You must give a textColor if button is not filled");
   final bool disabled;
   final bool hidden;
   final String content;
@@ -52,10 +56,10 @@ class HMButton extends HookWidget {
               ? const Color.fromRGBO(228, 229, 230, 1)
               : Color.alphaBlend(
                   buttonVariant == HMButtonVariant.filled
-                      ? Colors.black.withOpacity(isPressed ? 0.05 : 0.0)
+                      ? Colors.black.withOpacity(isPressed ? 0.1 : 0.0)
                       : textColor!.withOpacity(isPressed ? 0.1 : 0.0),
                   buttonVariant == HMButtonVariant.filled
-                      ? fillColor!
+                      ? fillColor ?? defaultColor
                       : Colors.transparent),
           border: buttonVariant == HMButtonVariant.outlined
               ? Border.all(
@@ -73,7 +77,7 @@ class HMButton extends HookWidget {
       buttonVariant == HMButtonVariant.filled
           ? child.padding(all: 3).decorated(
               border: isPressed
-                  ? Border.all(color: fillColor ?? Colors.transparent, width: 2)
+                  ? Border.all(color: fillColor ?? defaultColor, width: 2)
                   : null,
               borderRadius: BorderRadius.circular(radius * 1.2))
           : child;
@@ -97,9 +101,13 @@ class HMButton extends HookWidget {
         margin: EdgeInsets.symmetric(horizontal: getIconSize(size) / 5),
         child: Text(text)
             .fontSize(getTextSize(size))
-            .textColor(disabled
-                ? const Color.fromRGBO(173, 181, 189, 1.0)
-                : textColor!.withOpacity(isPressed ? 0.5 : 1.0))
+            .textColor(
+              disabled
+                  ? const Color.fromRGBO(173, 181, 189, 1.0)
+                  : buttonVariant == HMButtonVariant.filled
+                      ? textColor ?? checkColor(fillColor ?? defaultColor)
+                      : textColor!.withOpacity(isPressed ? 0.5 : 1.0),
+            )
             .alignment(Alignment.center),
       ),
     ];