+++ lib/components/select/hm_multi_select.dart
@@ -5,20 +5,20 @@ import 'package:styled_widget/styled_widget.dart';
 import '../../utils/constant.dart';
 import '../../size/hm_select_size.dart';
 
-class HMSelect extends HookWidget {
-  const HMSelect(
+class HMMultiSelect extends HookWidget {
+  const HMMultiSelect(
       {this.border,
       this.disabled = false,
       this.hidden = false,
-      this.value,
       this.boxRadius = 4,
+      this.multiSelectValue = const [],
       this.textColor = Colors.black,
       this.separatorLineColor,
       this.separatorLineHeight = 1,
       this.selectIcon = Icons.check,
-      this.selectList = const [],
+      required this.selectList,
       this.size = HMSelectSize.md,
-      this.selectIconColor,
+      this.iconColor,
       this.isLeft = false,
       required this.onChanged,
       super.key});
@@ -28,11 +28,11 @@ class HMSelect extends HookWidget {
   final HMSelectSize size;
   final double boxRadius;
   final IconData selectIcon;
-  final Color? selectIconColor;
+  final Color? iconColor;
   final Color textColor;
   final Color? separatorLineColor;
   final double separatorLineHeight;
-  final dynamic value;
+  final List<dynamic> multiSelectValue;
   final List selectList;
   final Border? border;
 
@@ -40,7 +40,7 @@ class HMSelect extends HookWidget {
   ///`"true"` to put the icon before the title
   ///and `"false"`to put the icon to end.
   final bool isLeft;
-  final void Function(dynamic value) onChanged;
+  final void Function(List value) onChanged;
 
   double _getTextSize(HMSelectSize size) {
     switch (size) {
@@ -63,32 +63,31 @@ class HMSelect extends HookWidget {
       Visibility(visible: !hidden, child: child);
 
   Widget _styledSelectPannel({
-    required ValueNotifier<dynamic> selection,
+    required ValueNotifier<List> multiSelection,
   }) {
-    return buildSelectItems(selection);
+    return buildMultiSelectItems(multiSelection);
   }
 
   @override
   Widget build(BuildContext context) {
-    final ValueNotifier selection = useState(value);
+    final ValueNotifier<List> multiSelection = useState(multiSelectValue);
 
     return AbsorbPointer(
         absorbing: disabled,
         child: _styledSelectPannel(
-          selection: selection,
+          multiSelection: multiSelection,
         ).parent(({required child}) => _styledBox(
               child: child,
             )));
   }
 
-  Widget buildSelectItems(ValueNotifier<dynamic> selection) {
+  Widget buildMultiSelectItems(ValueNotifier<List> selection) {
     return Container(
       decoration: BoxDecoration(
-          border: border ?? Border.all(color: Colors.black38),
+          border: border ?? Border.all(color: outlineColor),
           borderRadius: BorderRadius.circular(boxRadius)),
       child: ListView.separated(
         padding: EdgeInsets.zero,
-        // physics: BouncingScrollPhysics(),
         itemCount: selectList.length,
         shrinkWrap: true,
         separatorBuilder: (BuildContext context, int sIndex) {
@@ -97,11 +96,11 @@ class HMSelect extends HookWidget {
             indent: 0.0,
             endIndent: 0.0,
             thickness: separatorLineHeight,
-            // color: Colors.black,
+            color: separatorLineColor,
           );
         },
         itemBuilder: (BuildContext context, int index) {
-          final bool isSelected = selection.value == selectList[index];
+          final bool isSelected = selection.value.contains(selectList[index]);
           final List<Widget> children = [
             Text(
               '${selectList[index]}',
@@ -114,16 +113,24 @@ class HMSelect extends HookWidget {
               padding: const EdgeInsets.only(right: 20.0),
               child: Icon(
                 isSelected ? selectIcon : null,
-                color: disabled ? Colors.grey : selectIconColor ?? defaultColor,
+                color: disabled ? Colors.grey : iconColor ?? defaultColor,
                 size: _getTextSize(size) * 1.3,
               ),
             ),
           ];
           return GestureDetector(
             onTap: () {
-              if (selectList[index] != selection.value) {
-                selection.value = selectList[index];
-                // print('$index, ${selection.value}');
+              if (isSelected) {
+                final List x = List.from(selection.value);
+                x.remove(selectList[index]);
+                // print('remove from  x: $x');
+                selection.value = x;
+                onChanged(selection.value);
+              } else {
+                final List x = List.from(selection.value);
+                x.add(selectList[index]);
+                // print('This is x:  $x');
+                selection.value = x;
                 onChanged(selection.value);
               }
             },