+++ lib/type/hm_slider_type.dart
@@ -1,3 +1,9 @@
 //
 
 enum SliderType { simple, range }
+
+class HMSliderMark {
+  HMSliderMark({required this.value, required this.label});
+  final int value;
+  final String label;
+}