+++ lib/utils/constant.dart
@@ -1,3 +1,4 @@
 import 'package:flutter/material.dart';
 
 Color outlineColor = const Color.fromRGBO(206, 212, 218, 1);
+Color defaultColor = const Color.fromRGBO(121, 80, 242, 1);