+++ CHANGELOG.md
@@ -1,2 +1,12 @@
+## 0.0.2+1
+Fixed some bugs.
+Add hidden widget property.
+Add disable widget property.
+Display of the text color of the buttons according to the background.
+Update pubspec.yaml.
+Update README.md.
+## 0.0.2
+Move example file and update README.md
+Update pubspec.yaml
 ## 0.0.1
-First commit.
+Initial release.